moves functions from events.js to the only component that uses them
