+++ packages/vizbuilder/src/components/ConditionManager.jsx
@@ -3,12 +3,9 @@ import PropTypes from "prop-types";
 import {Button} from "@blueprintjs/core";
 import classNames from "classnames";
 
-import {
-  addCondition,
-  removeCondition,
-  updateCondition
-} from "../actions/events";
 import {fetchQuery} from "../actions/fetch";
+import OPERATORS from "../helpers/operators";
+import {makeRandomId} from "../helpers/random";
 
 import ConditionItem from "./ConditionItem";
 
@@ -19,9 +16,49 @@ class ConditionManager extends React.Component {
     super(props);
 
     this.fetchQuery = fetchQuery.bind(this);
-    this.addCondition = addCondition.bind(this);
-    this.updateCondition = updateCondition.bind(this);
-    this.removeCondition = removeCondition.bind(this);
+    this.addCondition = this.addCondition.bind(this);
+    this.updateCondition = this.updateCondition.bind(this);
+    this.removeCondition = this.removeCondition.bind(this);
+  }
+
+  addCondition() {
+    const {conditions} = this.props.query;
+    const {stateUpdate} = this.context;
+
+    const newConditions = [].concat(conditions, {
+      hash: makeRandomId(),
+      operator: OPERATORS.EQUAL,
+      property: "",
+      type: "cut",
+      values: []
+    });
+    return stateUpdate({query: {conditions: newConditions}});
+  }
+
+  updateCondition(condition) {
+    const {conditions} = this.props.query;
+    const {loadControl} = this.context;
+
+    const index = conditions.findIndex(cond => cond.hash === condition.hash);
+
+    if (index > -1) {
+      loadControl(() => {
+        const newConditions = conditions.slice();
+        newConditions.splice(index, 1, condition);
+        return {query: {conditions: newConditions}};
+      }, this.fetchQuery);
+    }
+  }
+
+  removeCondition(condition) {
+    const {conditions} = this.props.query;
+    const {loadControl} = this.context;
+
+    const newConditions = conditions.filter(cond => cond.hash !== condition.hash);
+
+    if (newConditions.length < conditions.length) {
+      loadControl(() => ({query: {conditions: newConditions}}), this.fetchQuery);
+    }
   }
 
   render() {
